Renamed the file
Just learning how to use github, so I dont know what I have done here lol. But would like to contribute to this repo. My implementation tries to do things in a similar way to the tutorial, and for this reason there is an entire section dedicated to a "library" similar to numpy. This keeps the layer and activation classes simple. It also helps keep main code that implement these classes clear and concise without any of the matrix calculation stuff getting in the way. Everything layer and activation objects do are also logged to the console displaying the result of the calculations at each step, which helps with learning. Logging can however, be disabled by setting debug macro to 0 if not needed.